chore(task9): add docs all test files and replace them to test_models from test_engine
